Fixed problems with constructors to super classes passing the wrong types.
Fixed a typo in the Start log.

+++ main/src/com/baptr/darkshaft/gfx/Wall.java
@@ -2,7 +2,7 @@ package com.baptr.darkshaft.gfx;
 
 public class Wall extends Defense {
 
-    public Wall(float x, float y, int facing) {
+    public Wall(int x, int y, int facing) {
         super(5, x, y);
     }
